add get_child_key_by name function
A function that can return a TcKey of a child associated with a String
is added. This requires the type checker to maintain a map of from
Strings to child indices.

Reason for the change is that types like structs do not operate with
children indices, but rather use names to identify a child, i.e. a
field.

Additionally, a renaming warning is fixed in `lib.rs`.